+++ src/main/java/org/recordy/server/external/controller/S3TestController.java
@@ -1,7 +1,7 @@
 package org.recordy.server.external.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.recordy.server.external.service.S3Service;
+import org.recordy.server.external.service.impl.S3ServiceImpl;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,18 +16,18 @@ import java.io.IOException;
 @RequestMapping("/api/v1")
 public class S3TestController implements S3TestApi{
 
-    private final S3Service s3Service;
+    private final S3ServiceImpl s3ServiceImpl;
     private static final String directoryPath = "recordy/file";
 
     @Override
     @PostMapping(path = "/uploadImageTest",consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public String uploadImageTest(@RequestPart MultipartFile image) throws IOException {
-        return this.s3Service.uploadImage(directoryPath,image);
+        return this.s3ServiceImpl.uploadImage(directoryPath,image);
     }
 
     @Override
     @PostMapping(path = "/uploadVideoTest",consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public String uploadVideoTest(@RequestPart MultipartFile video) throws IOException {
-        return this.s3Service.uploadVideo(directoryPath,video);
+        return this.s3ServiceImpl.uploadVideo(directoryPath,video);
     }
 }
