+++ src/main/java/org/recordy/server/external/service/S3ServiceImpl.java
@@ -1,0 +1,10 @@
+package org.recordy.server.external.service;
+
+public interface S3ServiceImpl {
+
+    //command
+
+
+    //query
+
+}
